feat: add review-app destroy hooks and a post-deploy hook
These new hooks allow users to further modify what commands happen during a deploy process, keeping the logic in simple scripts within the repository as opposed to in CI configuration. This ensures the ssh config does not need to be duplicated outside of the image in order to interact with Dokku apps during the release process.

Closes #115